requirements.txt
streamlit
pandas
langchain
langchain-community
langchain-groq
faiss-cpu
sentence-transformers
python-dotenv
